Update interest_points.cc
Solving warning when compiling with debug

+++ packages/imaging/interest_points/c_src/interest_points.cc
@@ -472,7 +472,7 @@ namespace InterestPoints {
   void SetPoints::addComponent() {
     if ((size_t) size != ccPoints->size())
         fprintf(stderr, "Size sincronization error %d %lu\n", size, ccPoints->size());
-    april_assert("Size sincronization error" && size == ccPoints->size());
+    april_assert("Size sincronization error" && (size_t)size == ccPoints->size());
     
     (*ccPoints).push_back(PointComponent());
     ++size;
